Extract LLM dialog code into LlmDialogHelper
Move prompt selector, custom prompt dialog, regenerate dialog, and
prompt execution logic from MainBibleActivity and BibleJavascriptInterface
into a dedicated LlmDialogHelper class to reduce bloat.
